api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
python-multipart>=0.0.6
joblib>=1.1.0
langchain>=0.1.0
langchain-community>=0.0.13
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.12.0
streamlit>=1.30.0
pypdf>=3.17.1
sse-starlette>=1.6.5
python-dotenv>=1.0.0 